refactor: extract _notify.py and _misc.py from helpers.py, inline single-callers
Create _notify.py (notification dispatch + response-size tracking) and
_misc.py (quota, hook-config, triage utilities) from helpers.py symbols.
Inline _import_and_call and _get_food_truck_prompt_builder into
tools_execution.py, and _build_hook_diagnostic_warning into
tools_kitchen.py. Redirect all passthrough re-exports to upstream
packages. Delete helpers.py.
