+++ src/autoskillit/server/tools_ci_watch.py
@@ -15,12 +15,9 @@ from autoskillit.core import CIRunScope, get_logger
 from autoskillit.pipeline import ToolContext
 from autoskillit.server import mcp
 from autoskillit.server._guards import _require_enabled
+from autoskillit.server._misc import resolve_repo_from_remote
+from autoskillit.server._notify import _notify, track_response_size
 from autoskillit.server._subprocess import _run_subprocess
-from autoskillit.server.helpers import (
-    _notify,
-    resolve_repo_from_remote,
-    track_response_size,
-)
 
 logger = get_logger(__name__)
