+++ src/autoskillit/server/tools_github.py
@@ -13,15 +13,12 @@ from fastmcp import Context
 from fastmcp.dependencies import CurrentContext
 
 from autoskillit.core import atomic_write, get_logger
+from autoskillit.execution import resolve_log_dir
 from autoskillit.pipeline import write_status
 from autoskillit.server import mcp
 from autoskillit.server._guards import _require_enabled
-from autoskillit.server.helpers import (
-    _extract_block,
-    _notify,
-    resolve_log_dir,
-    track_response_size,
-)
+from autoskillit.server._misc import _extract_block
+from autoskillit.server._notify import _notify, track_response_size
 
 if TYPE_CHECKING:
     from autoskillit.core import GitHubFetcher, HeadlessExecutor
