+++ src/autoskillit/server/tools_workspace.py
@@ -14,11 +14,8 @@ from fastmcp.dependencies import CurrentContext
 from autoskillit.core import get_logger, truncate_text
 from autoskillit.server import mcp
 from autoskillit.server._guards import _require_enabled
+from autoskillit.server._notify import _notify, track_response_size
 from autoskillit.server._subprocess import _run_subprocess
-from autoskillit.server.helpers import (
-    _notify,
-    track_response_size,
-)
 
 logger = get_logger(__name__)
