+++ tests/server/test_tools_execution_input_gates.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 import json
 
 import pytest
+from autoskillit.server.helpers import _get_config
 
 from autoskillit.server._guards import _check_dry_walkthrough
-from autoskillit.server.helpers import _get_config
 from autoskillit.server.tools_execution import run_skill
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
