+++ tests/server/test_tools_kitchen_gate.py
@@ -56,8 +56,9 @@ def test_hook_config_path_is_inside_temp_dir(tmp_path):
     from temp/.gitignore ('*') and can never produce a gitignore gap regardless of
     whatever entries are present in .gitignore or _AUTOSKILLIT_GITIGNORE_ENTRIES.
     """
-    from autoskillit.core.io import resolve_temp_dir
     from autoskillit.server.helpers import _hook_config_path
+
+    from autoskillit.core.io import resolve_temp_dir
 
     hook_path = _hook_config_path(tmp_path)
     temp_dir = resolve_temp_dir(tmp_path, None)
