+++ tests/server/test_tools_status_summaries.py
@@ -410,6 +410,7 @@ class TestGetLogRoot:
     def test_returns_resolved_log_dir(self, tool_ctx, tmp_path, monkeypatch):
         """_get_log_root() returns the resolved path for the configured log_dir."""
         from autoskillit.server.helpers import resolve_log_dir
+
         from autoskillit.server.tools_status import _get_log_root
 
         log_dir = tmp_path / "custom_logs"
